Edit AI Script update
update scripts, add ai
