+++ Assets/Script/Building/Player/P Resource Building.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 
 public class PResourceBuilding : MonoBehaviour
 {
+
     // Start is called before the first frame update
     void Start()
     {
