+++ Assets/Script/EnemyAI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e45b39fa2f362a4f8664c97aaf351d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
