--- Assets/Script/EnemyArrowContact.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EnemyArrowContact : MonoBehaviour
-{
-    private void OnTriggerEnter(Collider other)
-    {
-        if(other.gameObject.layer == LayerMask.NameToLayer("Player"))
-        {
-            Destroy(gameObject);
-        }
-        
-    }
-}
