--- Assets/Script/EnemyArrowContact.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: caf21914871d7f947a83501ad45197b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
