--- Assets/Script/PlayerArrowContact.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e165dc81d498402468956e8f6387e437
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
