--- Assets/Script/Shoot Arrow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: abeed9c5be9f4a144b2de9c831c846e9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
