--- Assets/Script/Turn Manager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4abfa0ebd887fcf40bc8f741f395887f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
