+++ Assets/Script/Unit/Active/overlapspere.cs
@@ -41,8 +41,8 @@ public class overlapspere : MonoBehaviour
     }
     private void OnDrawGizmos()
     {
-        //Gizmos.color = Color.red;
-        //Gizmos.DrawWireSphere(transform.position, radius);
+        Gizmos.color = Color.red;
+        Gizmos.DrawWireSphere(transform.position, radius);
 
         //Vector3 rightDirection = Quaternion.Euler(0, angle, 0) * transform.forward;
         //Gizmos.DrawRay(transform.position, rightDirection * distance);
