+++ Assets/prefab/Enemy Arrow.prefab
@@ -10,9 +10,9 @@ GameObject:
   m_Component:
   - component: {fileID: 6530645373980893404}
   - component: {fileID: -4926586706487383965}
-  - component: {fileID: -7209540009580813861}
   - component: {fileID: 467499413862053852}
   - component: {fileID: 6031040601198379549}
+  - component: {fileID: 2490138585655182676}
   m_Layer: 0
   m_Name: Enemy Arrow
   m_TagString: Untagged
@@ -50,18 +50,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 15
---- !u!114 &-7209540009580813861
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 626424037158238295}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: caf21914871d7f947a83501ad45197b5, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!65 &467499413862053852
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -110,6 +98,18 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
+--- !u!114 &2490138585655182676
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 626424037158238295}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ee0ea10971403064d84e60a3de256a29, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1741849185383052726
 GameObject:
   m_ObjectHideFlags: 0
