+++ Assets/prefab/Units/Enemy/Enemy Lancer.prefab
@@ -110,6 +110,10 @@ PrefabInstance:
         type: 3}
       insertIndex: -1
       addedObject: {fileID: -1301053320594332067}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 9d87951d064ffaa4c8ffe7e2408a0343,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 5564713845304669943}
   m_SourcePrefab: {fileID: 100100000, guid: 9d87951d064ffaa4c8ffe7e2408a0343, type: 3}
 --- !u!1 &80572098859112960 stripped
 GameObject:
@@ -207,9 +211,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   amount: 400
-  onDeath:
-    m_PersistentCalls:
-      m_Calls: []
 --- !u!114 &-1301053320594332067
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -231,6 +232,18 @@ MonoBehaviour:
   enemy_inrange: 0
   angle: 45
   distance: 3
+--- !u!114 &5564713845304669943
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 80572098859112960}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a8301b8f4ca103447a1b02159c9f8bce, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!4 &6332961455635280530 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 6502182839619065283, guid: 9d87951d064ffaa4c8ffe7e2408a0343,
