+++ Assets/prefab/Units/Player/Player Archer.prefab
@@ -99,6 +99,11 @@ PrefabInstance:
       propertyPath: m_Layer
       value: 8
       objectReference: {fileID: 0}
+    - target: {fileID: 919132149155446097, guid: 339769e19d1a3974a93f5fcdbe9fa01f,
+        type: 3}
+      propertyPath: m_TagString
+      value: Archer
+      objectReference: {fileID: 0}
     - target: {fileID: 5866666021909216657, guid: 339769e19d1a3974a93f5fcdbe9fa01f,
         type: 3}
       propertyPath: m_Controller
@@ -247,9 +252,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   amount: 100
-  onDeath:
-    m_PersistentCalls:
-      m_Calls: []
 --- !u!114 &3640204519741520043
 MonoBehaviour:
   m_ObjectHideFlags: 0
